+++ frontend/src/css/button.css
@@ -1,10 +1,15 @@
 .createButton{
     grid-area: herbList;
+    width: 150px;
+    padding: 5px 0px 10px 10px;
+    margin: 10px 50px;
+    margin-bottom: 10px;
     border-radius: 5px;
     background-color: #688662;
     cursor: pointer;
     color: #eeeeee;
-    align-items: end;
+    font-family: "Glacial Indifference";
+    font-size: 11pt;
 }
 
 input[type="file"] {
@@ -13,15 +18,17 @@ input[type="file"] {
 
 
 .seekPicture{
-    width: 150px;
+    width: 140px;
     padding: 5px 0px 10px 10px;
     margin: 0 50px;
     margin-bottom: 10px;
-
     border-radius: 5px;
     background-color: #688662;
     cursor: pointer;
     color: #eeeeee;
+    font-family: "Glacial Indifference";
+    font-size: 11pt;
+    text-align: center;
 }
 .uploadButton{
     grid-area: picture1;
@@ -33,6 +40,8 @@ input[type="file"] {
     background-color: #688662;
     cursor: pointer;
     color: #eeeeee;
+    font-family: "Glacial Indifference";
+    font-size: 11pt;
 }
 .saveButton{
     width: 150px;
@@ -43,6 +52,8 @@ input[type="file"] {
     background-color: #688662;
     cursor: pointer;
     color: #eeeeee;
+    font-family: "Glacial Indifference";
+    font-size: 11pt;
 }
 .deleteButton{
     width: 150px;
