+++ frontend/src/css/herbs.css
@@ -1,5 +1,7 @@
 .body{
     background-color: #f3f2f0;
+    font-family: "Glacial Indifference";
+    font-size: 13px;
 }
 
 .page {
@@ -50,12 +52,16 @@
     color: #688662;
 }
 .navBarSchrift{
+    padding: 20px 0px 0px 0px;
     font-family: "Glacial Indifference";
+    font-size: 12pt;
     color: #688662;
 }
 
 .herbList {
     grid-area: herbList;
+    font-family: "Glacial Indifference";
+    font-size: 12pt;
     position: center;
     padding-inline: 0px;
     padding: 15px ;
@@ -66,6 +72,8 @@
 }
 .herbListHeader {
     grid-area: herbList;
+    font-family: "Glacial Indifference";
+    font-size: 12pt;
     padding: 0px 0px 0px 50px;
     color: #d18c03;
 }
@@ -93,27 +101,27 @@
     grid-area: herbName;
     background-color: #e3dedb;
     font-family: "Glacial Indifference";
-    font-size: 30px;
+    font-size: 20pt;
     border: #eeeeee;
 }
 .herbDescription {
     grid-area: herbDescription;
     background-color: #e3dedb;
     font-family: "Glacial Indifference";
-    font-size: 15px;
+    font-size: 12pt;
 }
 .herbApplication {
     grid-area: herbApplication;
     background-color: #e3dedb;
     font-family: "Glacial Indifference";
-    font-size: 15px;
+    font-size: 12pt;
     border: none;
 }
 .herbNameCategory {
     grid-area: herbNameCategory;
     background-color: #f3f2f0;
     font-family: "Glacial Indifference";
-    font-size: 13px;
+    font-size: 12pt;
     margin-inside: 3px;
     padding: 30px 0px 0px 0px;
 }
@@ -121,14 +129,14 @@
     grid-area: herbDescriptionCategory;
     background-color: #f3f2f0;
     font-family: "Glacial Indifference";
-    font-size: 13px;
+    font-size: 12pt;
     border: none;
 }
 .herbApplicationCategory {
     grid-area: herbApplicationCategory;
     background-color: #f3f2f0;
     font-family: "Glacial Indifference";
-    font-size: 13px;
+    font-size: 12pt;
     border: none;
 
 }
